Docker: provides a way to test the playbook
Execute the Docker/README.md for instructions. The "retext"
command is needed. Or, just cat the file.

Dockerfiles will copy the relevant Ansible files to the target image
so tests can be run on it. Probably as important, changes can be
made to the packages selected by the local copy of the playbook and
a working version can then be pushed to github.

Initially, only the Ubuntu Dockerfiles have the instructions
that build MythTV and plugins. For now, they're commented out.