yay (pacman): download yay repo, add instructions
Probably the best solution for AUR packages due to issues with the yay
(and really makepkg) commands. They are documented in the yay.yml file
for future reference.

Download the yay repo, then tell the user what to do.

Thanks to Jim Abernathy for testing.

Also stopped downloading the Docker directory itself when running
Dockerfiles.

Refs: #67