User options: add user_options.yml and remove Docker
The user_options.yml file isn't under git and that's intended to allow
users to do some local configuration without it being overwritten by
the next git pull.

Docker is used just for testing and is no longer released with the repo.
It can be added back if there's interest.